requirements.txt
tensorflow==2.5.0
numpy~=1.19.5
scikit-learn~=0.24.1
matplotlib~=3.4.1
pandas~=1.2.3
spectral~=0.22.2
scipy~=1.6.2
uncertainty-calibration
pillow~=8.2.0
six~=1.15.0
setuptools~=52.0.0
tqdm~=4.60.0
sewar~=0.4.4
scikit-image~=0.14.2
pyyaml~=5.4.1
sklearn~=0.0
openpyxl~=3.0.7
pymoo~=0.4.2.2
jinja2~=3.0.1